application/src/test/java/application/storage/PersistanceStorageTest.java
package application.storage;

import static org.junit.jupiter.api.Assertions.assertEquals;

import java.io.IOException;

import org.jdom2.JDOMException;
import org.jmock.Expectations;
import org.jmock.imposters.ByteBuddyClassImposteriser;
import org.jmock.junit5.JUnit5Mockery;
import org.junit.jupiter.api.AfterEach;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.RegisterExtension;

import application.storage.xml.SystemTransformationsXMLFile;
import application.storage.xml.TaskDescriptionXMLFile;
import planning.method.SystemTransformations;
import planning.method.TaskDescription;

public class PersistanceStorageTest {

	@RegisterExtension
	JUnit5Mockery context = new JUnit5Mockery() {
		{
			setImposteriser(ByteBuddyClassImposteriser.INSTANCE);
		}
	};

	@AfterEach
	public void teardown() {
		context.assertIsSatisfied();
	}

	PersistanceStorage testable;

	SystemTransformationsXMLFile systemTransformationsXMLFile_mock;

	TaskDescriptionXMLFile taskDescriptionXMLFile_mock;

	@BeforeEach
	public void setup() {
		systemTransformationsXMLFile_mock = context.mock(SystemTransformationsXMLFile.class);
		taskDescriptionXMLFile_mock = context.mock(TaskDescriptionXMLFile.class);

		testable = new PersistanceStorage(systemTransformationsXMLFile_mock, taskDescriptionXMLFile_mock);
	}

	@Test
	public void newInstance() {
		testable = new PersistanceStorage();
	}

	@Test
	public void saveSystemTransformations() throws IOException {
		final SystemTransformations systemTransformations_mock = context.mock(SystemTransformations.class);
		final String path = "path-to-file";

		context.checking(new Expectations() {
			{
				oneOf(systemTransformationsXMLFile_mock).save(systemTransformations_mock, path);
			}
		});

		testable.saveSystemTransformations(systemTransformations_mock, path);
	}

	@Test
	public void saveTaskDescription() throws IOException {
		final TaskDescription taskDescription_mock = context.mock(TaskDescription.class);
		final String path = "path-to-file";

		context.checking(new Expectations() {
			{
				oneOf(taskDescriptionXMLFile_mock).save(taskDescription_mock, path);
			}
		});

		testable.saveTaskDescription(taskDescription_mock, path);
	}

	@Test
	public void loadSystemTransformations() throws IOException, JDOMException {
		final SystemTransformations systemTransformations_mock = context.mock(SystemTransformations.class);
		final String path = "path-to-file";

		context.checking(new Expectations() {
			{
				oneOf(systemTransformationsXMLFile_mock).load(path);
				will(returnValue(systemTransformations_mock));
			}
		});

		assertEquals(systemTransformations_mock, testable.loadSystemTransformations(path));
	}

	@Test
	public void loadTaskDescription() throws IOException, JDOMException {
		final TaskDescription taskDescription_mock = context.mock(TaskDescription.class);
		final String path = "path-to-file";

		context.checking(new Expectations() {
			{
				oneOf(taskDescriptionXMLFile_mock).load(path);
				will(returnValue(taskDescription_mock));
			}
		});

		assertEquals(taskDescription_mock, testable.loadTaskDescription(path));
	}
}
